Move Update component into Event folder
Change-type: patch
