+++ lib/Event/Update/Wrapper.jsx
@@ -4,7 +4,7 @@ import _ from 'lodash'
 import {
 	Box
 }	from 'rendition'
-import * as helpers from '../services/helpers'
+import * as helpers from '../../services/helpers'
 
 // Min-width is used to stop text from overflowing the flex container, see
 // https://css-tricks.com/flexbox-truncated-text/ for a nice explanation
