+++ lib/index.js
@@ -63,7 +63,7 @@ export {
 } from './Timeline'
 export {
 	default as Update
-} from './Update'
+} from './Event/Update'
 export {
 	UserAvatar, UserAvatarLive
 } from './UserAvatar'
